# lib/openapi/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timesheet API",
    "version": "1.0.0",
    "description": "Complete REST API for the Timesheet workforce management system. Handles employee time tracking, scheduling, roster management, payroll and analytics.\n\n## Authentication\nUses cookie-based JWT authentication. Three separate auth layers:\n- Admin/user: login via /api/auth/login\n- Employee: login via /api/employee/login\n- Device: register via /api/device/register\n\n## Roles\n- super_admin: full system access\n- admin: manage employees and schedules\n- user: limited view-only dashboard access\n- employee: clock in/out and own timesheets\n\n## Base URL\nAll endpoints are prefixed with /api\n\n## Error Responses\nAll errors follow the format: { error: string, message?: string }\n\n## Cookie Auth\nAll protected endpoints require valid JWT cookie sent automatically via credentials: include on all requests."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["error"]
      },
...
      "ObjectId": {
        "type": "string",
        "pattern": "^[0-9a-fA-F]{24}$",
        "example": "507f1f77bcf86cd799439011"
      },
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2024-03-07T10:30:00.000Z"
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user", "super_admin"],
            "example": "user"
          },
          "location": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "rights": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["read_employees", "write_timesheets"]
          },
          "managedRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current authenticated user profile",
        "description": "Returns the profile information for the currently authenticated admin or user. Includes role, permissions, assigned locations, and account details. Requires valid JWT session cookie. Used to verify authentication status and load user context in frontend applications.",
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/employee/upload/image": {
      "post": {
        "tags": ["Employee Clock"],
        "summary": "Upload employee clock-in photo or profile image",
        "description": "Uploads employee photos captured during clock-in operations or for profile updates. Supports image optimization, format validation, and secure storage via Cloudinary integration. Used for identity verification during time clock operations and maintaining current employee profile photos. Handles image compression and generates secure URLs for storage.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "imageId": {
                      "type": "string",
                      "example": "emp_img_123456789"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com/employee-images/emp_img_123456789.jpg"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or file too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/image": {
      "get": {
        "tags": ["Upload"],
        "summary": "Retrieve image file by unique identifier",
        "description": "Retrieves an image file from cloud storage using its unique identifier. Supports various image formats and automatic format optimization. Used for displaying employee profile photos, clock-in verification images, and other business-related images. Provides secure access to stored images with proper authentication.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "img_123456789"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image file",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Image not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}